ZF-5386: add safeguard for GC
git-svn-id: http://framework.zend.com/svn/framework/standard/trunk@14003 44c647ce-9c0f-0410-b52a-842ac1e357ba

+++ library/Zend/Captcha/Image.php
@@ -566,7 +566,12 @@ class Zend_Captcha_Image extends Zend_Captcha_Word
     protected function _gc()
     {
         $expire = time() - $this->getExpiration();
-        foreach (new DirectoryIterator($this->getImgDir()) as $file) {
+        $imgdir = $this->getImgDir();
+        if(!$imgdir || strlen($imgdir) < 2) {
+            // safety guard
+            return;
+        }
+        foreach (new DirectoryIterator($imgdir) as $file) {
             if (!$file->isDot() && !$file->isDir()) {
                 if ($file->getMTime() < $expire) {
                     unlink($file->getPathname());
